+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [unreleased]
 
+### Features
+- add support for better Buffer JSON stringify (#796)
+
 ### Fix
 - handle dom parser exception for invalid xml (#821)
 - do not log stack on assertions error (AnWeber/vscode-httpyac#338)